requirements.txt
Django
djangorestframework
psycopg2
drf-spectacular
python-dotenv
openai
PyPDF2
requests
faiss-cpu
sentence-transformers
langchain
djangorestframework-simplejwt